requirements.txt
streamlit
pandas
numpy
tensorflow
Pillow
matplotlib